transcription-agent/requirements.txt
livekit-agents~=0.12
livekit-plugins-deepgram~=0.7
python-dotenv~=1.0
